+++ dubber/BUILD
@@ -1,12 +1,13 @@
 load("@pip//:requirements.bzl", "requirement")
-load("@rules_python//python:defs.bzl", "py_binary")
+load("@rules_python//python:defs.bzl", "py_binary", "py_library")
 
 py_binary(
     name = "overlay",
     main = "overlay.py",
     srcs = ["src/overlay.py"],
     deps = [
-      requirement("pydub"),
+        ":common",
+        requirement("pydub"),
     ],
     visibility = ["//visibility:public"],
 )
@@ -16,7 +17,14 @@ py_binary(
     main = "concat.py",
     srcs = ["src/concat.py"],
     deps = [
-      requirement("pydub"),
+        ":common",
+        requirement("pydub"),
     ],
     visibility = ["//visibility:public"],
+)
+
+py_library(
+    name = "common",
+    srcs = glob(["src/common/**/*.py"]),
+    deps = [],
 )
